+++ web/components/index/KirakiraEffect.tsx
@@ -5,31 +5,10 @@ import { motion } from 'framer-motion';
 import KirakiraImage from '../../public/index/kirakira.png';
 
 import { KirakiraEffectProps } from '../../types/types_index'
+import { useMaxSize } from '../../hooks/hooks_index'
 
 const KirakiraEffect: React.FC<KirakiraEffectProps> = ({ effects }) => {
-  const [maxSize, setMaxSize] = useState({ maxWidth: '10vw', maxHeight: '10vw' });
-
-  useEffect(() => {
-    const handleResize = () => {
-      const screenWidth = window.innerWidth;
-      if (screenWidth > 1200) {
-        setMaxSize({ maxWidth: '35px', maxHeight: '35px' });  // 大きな画面の場合
-      } else if (screenWidth > 768) {
-        setMaxSize({ maxWidth: '25px', maxHeight: '25px' });  // 中くらいの画面の場合
-      } else {
-        setMaxSize({ maxWidth: '15px', maxHeight: '15px' });  // 小さな画面の場合
-      }
-    };
-
-    // 初回のサイズ設定
-    handleResize();
-
-    // リサイズイベントのリスナーを設定
-    window.addEventListener('resize', handleResize);
-
-    // クリーンアップ
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
+  const maxSize = useMaxSize(); 
   
   return (
     <>
