+++ web/hooks/hooks_index.ts
@@ -99,3 +99,31 @@ export const useTextVisibility = ({
     }
   }, [isHovered, currentImageIndex, hoverStrawberryImages, currentColor, onLastImage, onHoverEnd]);
 };
+
+export const useMaxSize = () => {
+  const [maxSize, setMaxSize] = useState({ maxWidth: '10vw', maxHeight: '10vw' });
+
+  useEffect(() => {
+    const handleResize = () => {
+      const screenWidth = window.innerWidth;
+      if (screenWidth > 1200) {
+        setMaxSize({ maxWidth: '35px', maxHeight: '35px' });  // 大きな画面の場合
+      } else if (screenWidth > 768) {
+        setMaxSize({ maxWidth: '25px', maxHeight: '25px' });  // 中くらいの画面の場合
+      } else {
+        setMaxSize({ maxWidth: '15px', maxHeight: '15px' });  // 小さな画面の場合
+      }
+    };
+
+    // 初回のサイズ設定
+    handleResize();
+
+    // リサイズイベントのリスナーを設定
+    window.addEventListener('resize', handleResize);
+
+    // クリーンアップ
+    return () => window.removeEventListener('resize', handleResize);
+  }, []);
+
+  return maxSize;
+};
